Move Caravel.Core public extensions methods to Caravel.Core.Extensions library.
