+++ examples/InMemoryJourneyDemo/InMemoryJourneyDemo.csproj
@@ -23,6 +23,7 @@
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\..\src\Caravel.Abstractions\Caravel.Abstractions.csproj" />
+        <ProjectReference Include="..\..\src\Caravel.Core.Extensions\Caravel.Core.Extensions.csproj" />
         <ProjectReference Include="..\..\src\Caravel.Core\Caravel.Core.csproj" />
         <ProjectReference Include="..\..\src\Caravel.Graph.Dijkstra\Caravel.Graph.Dijkstra.csproj" />
         <ProjectReference Include="..\..\src\Caravel.Mermaid\Caravel.Mermaid.csproj" />
