+++ examples/PlaywrightDemo/WebSite.Facade/WebSite.Facade.csproj
@@ -9,6 +9,7 @@
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\..\..\src\Caravel.Abstractions\Caravel.Abstractions.csproj" />
+        <ProjectReference Include="..\..\..\src\Caravel.Core.Extensions\Caravel.Core.Extensions.csproj" />
         <ProjectReference Include="..\..\..\src\Caravel.Core\Caravel.Core.csproj" />
         <ProjectReference Include="..\..\..\src\Caravel.Graph.Dijkstra\Caravel.Graph.Dijkstra.csproj" />
     </ItemGroup>
