+++ src/Caravel.Core/Extensions/JourneyExtensions.cs
@@ -4,116 +4,17 @@ namespace Caravel.Core.Extensions;
 
 public static partial class JourneyExtensions
 {
-    public static async Task<IJourney> GotoAsync<TDestination>(
+    internal static CancellationTokenSource LinkJourneyAndLocalCancellationTokens(
         this IJourney journey,
-        CancellationToken scopedCancellationToken = default
+        CancellationToken scopedCancellationToken
     )
-        where TDestination : INode
     {
         ArgumentNullException.ThrowIfNull(journey, nameof(journey));
+        var linkedCancellationTokenSource = CancellationTokenSource.CreateLinkedTokenSource(
+            journey.JourneyCancellationToken,
+            scopedCancellationToken
+        );
 
-        return await journey
-            .GotoAsync<TDestination>(
-                Waypoints.Empty(),
-                ExcludedNodes.Empty(),
-                scopedCancellationToken
-            )
-            .ConfigureAwait(false);
-    }
-
-    public static async Task<IJourney> GotoAsync<TDestination>(
-        this IJourney journey,
-        ExcludedNodes excludedNodes,
-        CancellationToken scopedCancellationToken = default
-    )
-        where TDestination : INode
-    {
-        ArgumentNullException.ThrowIfNull(journey, nameof(journey));
-        return await journey
-            .GotoAsync<TDestination>(Waypoints.Empty(), excludedNodes, scopedCancellationToken)
-            .ConfigureAwait(false);
-    }
-
-    public static async Task<IJourney> GotoAsync<TDestination>(
-        this IJourney journey,
-        Waypoints waypoints,
-        CancellationToken scopedCancellationToken = default
-    )
-        where TDestination : INode
-    {
-        ArgumentNullException.ThrowIfNull(journey, nameof(journey));
-        return await journey
-            .GotoAsync<TDestination>(waypoints, ExcludedNodes.Empty(), scopedCancellationToken)
-            .ConfigureAwait(false);
-    }
-
-    public static async Task<IJourney> GotoAsync<TDestination>(
-        this IJourney journey,
-        Waypoints waypoints,
-        ExcludedNodes excludeNodes,
-        CancellationToken scopedCancellationToken = default
-    )
-        where TDestination : INode
-    {
-        ArgumentNullException.ThrowIfNull(journey, nameof(journey));
-        return await journey
-            .GotoAsync<TDestination>(waypoints, excludeNodes, scopedCancellationToken)
-            .ConfigureAwait(false);
-    }
-
-    public static async Task<IJourney> DoAsync<TCurrentNode>(
-        this IJourney journey,
-        Func<TCurrentNode, CancellationToken, Task<TCurrentNode>> func,
-        CancellationToken scopedCancellationToken = default
-    )
-        where TCurrentNode : INode
-    {
-        ArgumentNullException.ThrowIfNull(journey, nameof(journey));
-        return await journey
-            .DoAsync<TCurrentNode, TCurrentNode>(
-                (_, node, token) => func(node, token),
-                scopedCancellationToken
-            )
-            .ConfigureAwait(false);
-    }
-
-    public static async Task<IJourney> DoAsync<TCurrentNode, TNodeOut>(
-        this IJourney journey,
-        Func<TCurrentNode, CancellationToken, Task<TNodeOut>> func,
-        CancellationToken scopedCancellationToken = default
-    )
-        where TCurrentNode : INode
-        where TNodeOut : INode
-    {
-        ArgumentNullException.ThrowIfNull(journey, nameof(journey));
-        return await journey
-            .DoAsync<TCurrentNode, TNodeOut>(
-                (_, node, token) => func(node, token),
-                scopedCancellationToken
-            )
-            .ConfigureAwait(false);
-    }
-
-    public static async Task<IJourney> DoAsync<TCurrentNode>(
-        this IJourney journey,
-        Func<IJourney, TCurrentNode, CancellationToken, Task<TCurrentNode>> func,
-        CancellationToken scopedCancellationToken = default
-    )
-        where TCurrentNode : INode
-    {
-        ArgumentNullException.ThrowIfNull(journey, nameof(journey));
-        return await journey.DoAsync(func, scopedCancellationToken).ConfigureAwait(false);
-    }
-
-    public static async Task<IJourney> DoAsync<TCurrentNode, TNodeOut>(
-        this IJourney journey,
-        Func<IJourney, TCurrentNode, CancellationToken, Task<TNodeOut>> func,
-        CancellationToken scopedCancellationToken = default
-    )
-        where TCurrentNode : INode
-        where TNodeOut : INode
-    {
-        ArgumentNullException.ThrowIfNull(journey, nameof(journey));
-        return await journey.DoAsync(func, scopedCancellationToken).ConfigureAwait(false);
+        return linkedCancellationTokenSource;
     }
 }
